create flatteningMappingTemplatesAndDataSources and used when configuration have to be loaded from file or has been passed as parameter
